requirements.txt
psycopg2-binary
SQLAlchemy~=2.0.43
alembic
pytest~=8.4.2
pytest-asyncio
python-telegram-bot~=22.5
python-dotenv~=1.1.1
pydantic-settings~=2.11.0
requests~=2.32.5
httpx~=0.27.2
fastapi~=0.115.2
uvicorn~=0.29.0
beautifulsoup4~=4.14.2
selenium~=4.38.0
webdriver-manager~=4.0.2
pandas~=2.3.3
undetected-chromedriver

torch~=2.8.0+cu126
unsloth~=2025.11.1
sentence-transformers~=5.1.2

pgvector~=0.4.1
numpy~=2.3.4
faiss-cpu~=1.9.0
hdbscan~=0.8.40
trl~=0.23.0
packaging~=25.0
transformers~=4.57.1
